refactor: moving iterator stream into its own file
